bin: List kube_control_plane + kube_node in labels and taints
In Terraform generated inventories k8s_cluster is rendered with each
host individually. However, a common use case is that a static inventory
has:

[k8s_cluster:children]
kube_control_plane
kube_node

This use case was not supported.

This is now fixed by listing the hosts in the groups kube_control_plane
and kube_node instead which should have the hosts listed individually.